feat: move user implementation to shared module
