--- app/src/main/java/com/singaludra/gofood/feature/user/domain/SaveUserInfo.kt
@@ -1,7 +1,0 @@
-package com.singaludra.gofood.feature.user.domain
-
-import com.singaludra.shared.domain.UserRoot
-
-interface SaveUserInfo {
-    suspend fun save(userRoot: com.singaludra.shared.domain.UserRoot)
-}
