--- app/src/main/java/com/singaludra/gofood/feature/user/local/IUserInfoRepository.kt
@@ -1,7 +1,0 @@
-package com.singaludra.gofood.feature.user.local
-
-import com.singaludra.gofood.feature.user.domain.UserEntity
-
-interface IUserInfoRepository {
-    suspend fun save(userEntity: UserEntity)
-}
