+++ app/src/main/java/com/singaludra/gofood/main/decorators/RegisterUserDecorator.kt
@@ -3,7 +3,7 @@ package com.singaludra.gofood.main.decorators
 import com.singaludra.register.domain.RegisterUser
 import com.singaludra.register.domain.RegisterUserResult
 import com.singaludra.register.domain.request.UserData
-import com.singaludra.gofood.feature.user.domain.SaveUserInfo
+import com.singaludra.shared.user.domain.SaveUserInfo
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.flow
