+++ app/src/main/java/com/singaludra/gofood/main/factories/register/RegisterUserDecoratorFactory.kt
@@ -1,7 +1,7 @@
 package com.singaludra.gofood.main.factories.register
 
 import com.singaludra.register.domain.RegisterUser
-import com.singaludra.gofood.feature.user.domain.SaveUserInfo
+import com.singaludra.shared.user.domain.SaveUserInfo
 import com.singaludra.gofood.main.decorators.RegisterUserDecorator
 
 class RegisterUserDecoratorFactory {
