+++ shared/src/main/java/com/singaludra/shared/ui/theme/Color.kt
@@ -10,4 +10,4 @@ val Purple40 = Color(0xFF6650a4)
 val PurpleGrey40 = Color(0xFF625b71)
 val Pink40 = Color(0xFF7D5260)
 
-val Green = Color(0xFF63A546)
+val Green = Color(0xFF63A555)
