+++ shared/src/main/java/com/singaludra/shared/user/domain/SaveUserInfo.kt
@@ -1,0 +1,7 @@
+package com.singaludra.shared.user.domain
+
+import com.singaludra.shared.domain.UserRoot
+
+interface SaveUserInfo {
+    suspend fun save(userRoot: UserRoot)
+}
