+++ shared/src/main/java/com/singaludra/shared/user/local/IUserInfoRepository.kt
@@ -1,0 +1,7 @@
+package com.singaludra.shared.user.local
+
+import com.singaludra.shared.user.domain.UserEntity
+
+interface IUserInfoRepository {
+    suspend fun save(userEntity: UserEntity)
+}
